LayoutTests/media/video-interruption-with-resume-not-allowing-play-expected.txt


Test that playback is paused by an interruption, and that ending the interruption does automatically resume playback.

EVENT(canplaythrough)
RUN(video.play())

EVENT(playing)
RUN(internals.beginMediaSessionInterruption())

100ms timer fired...
EXPECTED (video.paused == 'true') OK
RUN(internals.endMediaSessionInterruption(''))

100ms timer fired...
EXPECTED (video.paused == 'true') OK

END OF TEST
